# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NusaAlert API",
        "description": "API dokumentasi untuk NusaAlert.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/v1/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Statistik dashboard admin",
                "operationId": "801e40951e675fc46835d28c630c3a07",
                "responses": {
                    "200": {
                        "description": "Data statistik admin",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Akses admin diperlukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/laporan/{id}/verify": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Verifikasi laporan komunitas",
                "operationId": "7c109ac6ffa181813b07ca78841937b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Laporan berhasil diverifikasi"
                    },
                    "403": {
                        "description": "Akses admin diperlukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/laporan/{id}/reject": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Tolak dan hapus laporan komunitas",
                "operationId": "f230911bc79e1fba9f5071f1b0bf54b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Laporan berhasil ditolak"
                    },
                    "403": {
                        "description": "Akses admin diperlukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/users/{id}/role": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update role user",
                "operationId": "31dfe3c5aebc71c1f7a33106a8c9ad33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "role"
                                ],
                                "properties": {
                                    "role": {
                                        "type": "string",
                                        "example": "member"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role user berhasil diperbarui"
                    },
                    "403": {
                        "description": "Akses admin diperlukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/bencana": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Tambah data bencana manual",
                "operationId": "5e650fceda8493bc2627eb9fcebbdb11",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "jenis_bencana",
                                    "latitude",
                                    "longitude",
                                    "wilayah"
                                ],
                                "properties": {
                                    "jenis_bencana": {
                                        "type": "string",
                                        "enum": [
                                            "gempa",
                                            "tsunami",
                                            "banjir",
                                            "cuaca_ekstrem",
                                            "gunung_api",
                                            "tanah_longsor"
                                        ]
                                    },
                                    "magnitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5.2
                                    },
                                    "kedalaman_km": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -6.2088
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 106.8456
                                    },
                                    "wilayah": {
                                        "type": "string",
                                        "example": "Jakarta Selatan"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Gempa terasa kuat"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bencana berhasil ditambahkan"
                    },
                    "403": {
                        "description": "Akses admin diperlukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/bencana/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Hapus data bencana",
                "operationId": "2b59e95a6bda6458b3410227b91fef57",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bencana berhasil dihapus"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/laporan/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Hapus laporan komunitas",
                "operationId": "deb492c0cb6453485820fca619efa445",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Laporan berhasil dihapus"
                    },
                    "403": {
                        "description": "Akses admin diperlukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Daftar semua user",
                "operationId": "bb0ce6c803aec2e06a5a19aa0022bcd1",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar user berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "apiKey": {
                "type": "apiKey",
                "name": "X-API-Key",
                "in": "header"
            },
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    },
...
}